requirements.txt
astroid==2.8.0
attrs==19.3.0
autopep8==1.5.7
black==21.12b0
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
debugpy==1.5.1
docker==5.0.3
docutils==0.15.2
flake8==3.9.2
Flask==1.1.2
Flask-SocketIO==4.3.1
idna==2.10
importlib-metadata==1.7.0
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mercurial==5.9.1
more-itertools==8.4.0
multipledispatch==0.6.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.9.0
platformdirs==2.3.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.7.0
pyflakes==2.3.0
pylint==2.11.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.2.5
pytest-flake8==1.0.6
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.2
tomli==1.2.3
typing-extensions==3.10.0.2
urllib3==1.25.9
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
